requirements.txt
google-genai
python-dotenv
jsonschema
